third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/activation/focus-updated-style.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: focus on styled element</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">

<!--
Focus on a div that has updated style/layout.
-->

<div id="container">
  <div id="focusable" tabIndex="0">
    focusable thing
  </div>
</div>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>

<script>
promise_test((t) => {
  const acquirePromise = setInvisibleActivatable(container);
  return new Promise((resolve, reject) => {
    acquirePromise.then(() => {
      container.onbeforeactivate = reject;
      focusable.style = "display: none";
      focusable.focus();
      t.step(() => assert_not_equals(document.activeElement, focusable));
      resolve();
    });
  });
}, "Trying to focus on an element in a locked subtree that's not visible anymore won't work");
</script>
</html>
